Fix refactored imports
